requirements.txt
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.0
github3.py==3.2.0
idna==3.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-redmine==2.3.0
requests==2.28.1
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
